refactor: move Cranelift policy to backend
